// SGUEES-SPA/src/app/pages/General/gen-empresa/models/gen-empresa.ts
export interface GenEmpresa {
	CORR_EMPRESA: number;
	NOMBRE_EMPRESA: string;
	NOMBRE_COMERCIAL: string;
	NOMBRE_REPRESENTANTE_LEGAL: string;
	GIRO_EMPRESA: string;
	DIRECCION_EMPRESA: string;
	NUMERO_NIT: string;
	NUMERO_NRC: string;
	NOMBRE_CONTACTO: string;
	TELEFONO_1: string;
	TELEFONO_2: string;
	FAX: string;
	CORREO_ELECTRONICO: string;
	LOGO_1: string | null;
	LOGO_2: string | null;
	TAMANO_EMPRESA: string;
	NATURAL_JURIDICO: string;
	CODIGO_EMPRESA: string;
	CORR_PAIS: number;
	CORR_DEPTO: number;
	CORR_MUNICIPIO: number;
	NOMBRE_EMPRESA_LARGO: string;
	DIRECCION_EMPRESA_LARGO: string;
	SELLO: string | null;
	CODIGO_POSTAL: string;
	TIPO_INGRESO_ISR: number;
	CORR_SECTOR_ECONOMICO: number;
	USA_CAMPOS_LIBRO_IVA: boolean;
	PERMITE_EDITAR_CAMPOS_LIBRO_IVA: boolean;
	USUARIO_CREA: string;
	FECHA_CREA: Date;
	ESTACION_CREA: string;
	USUARIO_ACTU: string;
	FECHA_ACTU: Date;
	ESTACION_ACTU: string;
	NOMBRE_PAIS: string;
	NOMBRE_DEPTO: string;
	NOMBRE_MUNICIPIO: string;
	NOMBRE_SECTOR_ECONOMICO: string;
}